+++ index.html
@@ -6,6 +6,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="stylesheet" href="leaflet/leaflet.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
+  <script src="js/disable-form.js" type="module"></script>
   <title>Кексобукинг</title>
 </head>
 <body>
@@ -228,7 +229,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="js/main.js" type="module"></script>
   <script src="leaflet/leaflet.js"></script>
-</body>
+  <script src="js/main.js" type="module"></script>
 </html>
