+++ js/create-map.js
@@ -3,42 +3,10 @@
 import {similarArrays} from './data.js';
 import {FLOAT_COUNT, CENTER_TOKYO_LAT, CENTER_TOKYO_LNG, PIN_WIDTH, PIN_HEIGHT, SCALE_MAP} from './const.js';
 import {createCard} from './similar-card.js';
+import {activateForm} from './activate-form.js';
 
 const mapCanvas = document.querySelector('#map-canvas');
 const adFormAddress = document.querySelector('#address');
-const adForm = document.querySelector('.ad-form');
-const adFieldsetForm = adForm.querySelectorAll('fieldset');
-const mapFilters = document.querySelector('.map__filters');
-const mapFieldset = mapFilters.querySelector('fieldset');
-const mapSelect = mapFilters.querySelectorAll('select');
-
-const disableForm = function () {
-  adForm.classList.add('ad-form--disabled');
-  adFieldsetForm.forEach((value) => {
-    value.setAttribute('disabled', 'disabled');
-  });
-  mapFilters.classList.add('ad-form--disabled');
-  mapFieldset.setAttribute('disabled', 'disabled');
-  mapSelect.forEach((value) => {
-    value.setAttribute('disabled', 'disabled');
-  });
-
-};
-
-disableForm();
-
-const activateForm = function () {
-  adForm.classList.remove('ad-form--disabled');
-  adFieldsetForm.forEach((value) => {
-    value.removeAttribute('disabled', 'disabled');
-  });
-
-  mapFilters.classList.remove('ad-form--disabled');
-  mapFieldset.removeAttribute('disabled', 'disabled');
-  mapSelect.forEach((value) => {
-    value.removeAttribute('disabled', 'disabled');
-  });
-};
 
 //Создаем карту
 const map = L.map(mapCanvas)
